requirements.txt
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
blinker==1.9.0
boto3==1.39.4
botocore==1.39.4
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.2.18
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
limits==5.4.0
MarkupSafe==3.0.2
packaging==25.0
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
s3transfer==0.13.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
